asterisk 13: remove xivo_workaround_remote_bridge_crash
Bridging in asterisk 13 use the Bridging API and functions like
remote_bridge_loop and local_bridge_loop doesn't exist anymore. The
functionaly equivalent code is located in bridges/bridge_native_rtp.c.

I don't know if the race condition the patch was preventing can still
happens in asterisk 13; I need to understand better the new bridging API
and how masquerade occurs when a channel is a bridge. So this patch
might (but hopefully not) come back with a different implementation
later on.