AgentComplete event now always have a Uniqueid key
See https://issues.asterisk.org/jira/browse/ASTERISK-25187 for the bug
it fixes.

This is in fact only a partial patch for the bug, i.e. it only adds the
"Uniqueid" key and does not add the "DestChannel" key, because we only
needed the Uniqueid key.

As a side effect, it also adds a "Ast11Uniqueid" key to the
AgentComplete message.