QueueCallerLeave event is now always sent
Even when the caller's channel has been masqueraded into a Surrogate
channel.

See https://issues.asterisk.org/jira/browse/ASTERISK-25203 for the bug
it fixes.

The patch fix the problem for the QueueCallerEvent, that said there's
the same problem for the QueueAbandon event (which we are not using in
XiVO, so I did not fix it) and it could also theoritically happens for
the QueueCallerJoin event I believe.