app_queue: call update_queue before send_agent_complete
The base problem is that when a call between a caller and a queue member
is ended (either because of a hang up or transfer), the first
QueueMemberStatus AMI event that will be eventually sent will not have
its "LastCall" and "CallsTaken" value updated. An AgentComplete AMI
event will also be sent. In xivo-ctid, when we receive the AgentComplete
event, we ask for the queue member status of that member to get the
updated values. But in somes cases, the queue member state still hasn't
been updated.

With this patch, we make sure the queue member is updated *before* the
AgentComplete event is sent.