add xivo_voicemail_email_on_reload patch
Patch for #6114 (upstream bug ASTERISK-24463). Always make a copy of
email / emailbody / emailsubject when creating a copy of a struct
ast_vm_user in the find_user function.

That way, if a "voicemail reload / module reload app_voicemail" happens
when we have a reference to a copy of a VM user (because someone is
leaving a message in its voicemail), then the copy's reference to email
will still be valid after the reload, i.e. after that the reload thread
has freed memory associated to the original struct ast_vm_user.