don't manipulate potentially destroyed channel
After ast_bridge_impart returns, the thread can't reference the peer
channel since ast_bridge_impart took ownership of it and gave the
reference to another thread: when that thread terminates, the reference
count is decremented, potentially destroying the channel. And if you try
to clear a flag on a destroyed channel, you are playing with memory you
don't know belongs to what, which is causing all kind of nasty problems,
like crashes.

Fix the problem by not clearing the AST_FLAG_END_DTMF_ONLY flag. Not
clearing the flag is not an issue since the channel is usually hung up
at this point (see ast_bridge_impart).

Also, set and clear the flag on the channels while the channel lock is
held, to be on the safe side.

While working on this issue, I also realized that the patch doesn't
cover scenarios with Local channels. For example, if a SIP channel calls
another SIP channels via Local channels, and the SIP channels use SIP
INFO DTMF and direct media is enabled, then DTMF won't work.