nova: fix name on forward
When a call was forwarded to an outgoing call, the name for the outgoing
clal was the original extension called (i.e. the extension that
triggered the forward) instead of the extension that was called due to
the forward.

Also add "type" information in the names, i.e. is it a channel linked to
an incall/outcall/internal.