nova: fix name of channels created by local channels
The typical scenario that wasn't working is "Alice creates an outgoing
call, than transfer with *2 to Bob". Bob's channel name was seen as
outgoing instead of internal.

Also change how local channels name are handled: instead of having
mostly meaningless name, the name of a local channel is the name of the
channel it is proxying. For example, in the case:

   SIP/A-00 -> Local/s@foo-00;1 <-> Local/s@foo-00;2 -> SIP/B-00

Then the ;1 local channel name is the name of "SIP/B-00", and the ;2
local channel name is the name of "SIP/A-00".