add xivo_dtls_negotiation_delay patch
A patch similar to this one was added in the upstream asterisk 13 source
code in december 2015 to fix a "DTLS negotiation delays" isssue
(ASTERISK-25614) but was reverted in january 2016 because it was
likely causing a deadlock in certain scenarios (see ASTERISK-25645).

The commits that added the patch are in the upstream code are:

* 24ae124e4f7310cfa64c187b944b2ffc060da28d
* 965a0eee46d24321f74c244e23c5a5f45e67e12b
* 1e24a0ca8ae075af2814668fc99ecfabb47423b3

The patch added in this commit is slightly different then the original
patch because it doesn't add the ast_mutex_lock/ast_mutex_unlock in the
recvfrom function, because they were deemed not truly necessary.

Ticket #6358.