add fix for ASTERISK-26299
ASTERISK-26299: app_queue: Queue application sometimes stops calling
members with Local interface.

This actually reverts the fix for ASTERISK-16115 which was added in
asterisk upstream code in 13.10, since it is this fix that is causing the
bug described in ASTERISK-26299.

ASTERISK-16115: problem with ringinuse=no, queue members receive
sometimes two calls

The patch "xivo_revert_queue_members_can_receive_multiple_calls" will be
removed when ASTERISK-26299 is going to be fixed.