dockerfile: use asterisk debian package
Two main reasons to do this:

* reduce the size of the image (previously ~1.4 GB, now ~270 MB)
* reduce the time the image takes to build

This has some disadvantages, like not being able to build an image for a
version X other than the one on the xivo mirror, etc, but depending on
what's the goal of the image (is it to have an asterisk as close as the
one you'll usually find on a xivo?), then it might not matter that much.

This also remove xivo-confgend-client and asterisk configuration files from
the image.