xivo_skill_queues: fix 13 to 14 upgrade issue
In Asterisk 13, when a queue member was dumped to the asterisk DB, the
format was:

   ...;skills

In Asterisk 14, before this commit, it was:

   ...;reason_paused;skills

Thus, when upgrading Asterisk from 13 to 14 with agents logged in, when
app_queue would re-read the persistent queue members, it would think
that some members have no skills and a weird "reason paused" (corresponding
to the member skills). With this commit, it fix that issue by using the
format:

   ...;skills;reason_paused

So during an 13 to 14 upgrade (and even a 14 to 13 downgrade), the queue
members are loaded properly in respect of skills and reason paused.