add patch xivo_ari_expose_voicemail_functions
This patch allow the ast_vm_msg_remove and ast_vm_msg_move functions to
be used from ARI, in a pratical but not-so REST way.

Most of this patch is just boilerplate automatically generated when
running "make ari-stubs". Don't forget to rerun "make ari-stubs" if you
ever modify the swagger specification.