use bundled version of pjproject
This is actually the Asterisk preferred way to include pjproject.
See: https://wiki.asterisk.org/wiki/display/AST/Building+and+Installing+pjproject

There's quite a few advantages for doing this, but what pushed me to do it
was that I wanted to increase the maximum number of ICE candidates (and
ICE checks) from the pjproject default, and Asterisk is already doing
such thing in their bundled pjproject version.