restore realtime priority
Why:

* Since we moved to systemd, asterisk has lost its realtime priority
(-11). It may cause a loss of voice quality if some other process is
taking too much CPU.

How:

* To ask realtime priority, asterisk must run as root.
* It will then drop its privileges and switch to user asterisk as soon
as it can.
* The RuntimeDirectory would instruct systemd to create /run/asterisk
with owner root, blocking asterisk to create any file in there, such as
its control socket used for `asterisk -r`. So we have to create and
delete this directory manually.
* The directories created in ExecStartPre come from the removed init
file.