Fix TTL zero stopping recurring STUN resolution
Why:

* If the TTL received from the DNS server is zero, meaning the
cache has expired, the recurring DNS would terminate itself.
With this fix, if a TTL of zero is received, a default value of
30 second is applied to the next query.

This 30 second delay avoids spamming the DNS server unnecessarily
if received TTL is always zero.

This change addresses the need by:

* If STUN server resolution stops, and the server is down, every
connection in Asterisk will timeout for 10 seconds while trying to
reach the STUN server, severely degrading user experience.