close stream on error playing beep
Why:

* If channel is hungup before the beep has played, the file descriptor
will stay stuck opened.

This happens when a user does not stop recording before hanging up.  The beep file will be loaded
and the stream opened but nothing will happen and no cleanup afterwards.

This patch makes sure to close the stream (and the file) if the beep couldn't be played.