res_musiconhold: remove invalid moh class when it cannot playback
Why: if an invalid WAV file is provided in a MOH class, it causes
asterisk to stop playback of the music. In the case of a call transfer,
people may hear eachother when it is unwanted. Call holding also does
not work because the music never plays.

This patch fixes the issue by removing the culprit MOH class. However,
it only works after the first occurence of the issue. The first time the
MOH is played will still present the problematic behavior.